Move Resolve into TeePee assembly.
